requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Deprecated==1.2.18
distro==1.9.0
Django==5.2.1
django-allauth==65.9.0
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
groq==0.26.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpath-python==1.0.6
Markdown==3.8
openai==1.84.0
packaging==25.0
polar-sdk==0.20.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
standardwebhooks==1.0.0
tqdm==4.67.1
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.2.3
wrapt==1.17.2
yt-dlp==2024.12.13
